Seperate inflector from str::class
